Refactor salary calculator and cleanup the code
The major change introduced in this commit is an organised restructuring of the salary calculation logic in the App.kt file. Instead of bloated and somewhat redundant control flow when calculating salary based on hourly, daily, weekly etc. rates, now we utilize a generic function getSalaryResults which streamlines this process and makes the code cleaner, more maintainable and more understandable. Minor changes are also introduced to add some import libraries to support newly introduced features.